add yaml representation
The YamlRepresentation shares a lot of code with JsonRepresentation,
because they both use Jackson's ObjectMapper under the covers.  As a
result, I've extracted a JacksonRepresentation class with the common
functionality of both.
